+++ carrier/src/main/java/com/markchan/carrier/domain/Font.java
@@ -5,7 +5,7 @@ package com.markchan.carrier.domain;
  */
 public class Font {
 
-    private int id;
+    private Long id;
     private String displayName;
     private String postscriptName;
     private String thumbUrl;
@@ -14,7 +14,8 @@ public class Font {
     public Font() {
     }
 
-    public Font(int id, String displayName, String postscriptName, String thumbUrl, String uri) {
+    public Font(Long id, String displayName, String postscriptName, String thumbUrl,
+            String uri) {
         this.id = id;
         this.displayName = displayName;
         this.postscriptName = postscriptName;
@@ -22,11 +23,11 @@ public class Font {
         this.uri = uri;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
